put adaptors in own module + use bevy_reflect::Reflect directly
